Almost finished implementing the waves. Added new buttons to the GUI.

+++ FinalYearProject/TowerDefense/res/materials/shaders/transparentshader.frag
@@ -1,6 +1,6 @@
 
 uniform sampler2D myTexture;
-uniform double alpha;
+uniform float alpha;
 
 void main (void)
 {
